Added Bluetooth menu and new main screen layout
There is a new screen layout for the bluetooth menu where the majority
of the bluetooth code should be; i.e. where it scans for bluetooth
devices, and where the bluetooth car device should be chosen and paired.
Changed the main screen and added buttons. Events for each button has
been created, but the ones that transmit data through bluetooth will
need to be updated. I also changed the name of the new activity(Android
studio automatically changes the references to the new name anywhere
that it is used and needed, including the manifest).